+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Adds `Installer` property to configuration file for dynamically creating the ChocolateyInstall.ps1 file
+- Updates Chrome example to use new `Installer` property
+
 ## [0.0.6] - 2021-04-29
 
 ### Added